Feature/graphannis handles multiple corpora

+++ annis-gui/src/main/java/annis/gui/resultview/ResultViewPanel.java
@@ -418,7 +418,7 @@ public class ResultViewPanel extends VerticalLayout implements OnLoadCallbackExt
         return result;
     }
 
-    public void setCount(int count) {
+    public void setCount(long count) {
         paging.setCount(count, false);
         paging.setStartNumber(initialQuery.getOffset());
     }
